% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{csl_Flora}
\alias{csl_Flora}
\title{Descripción Flora}
\usage{
csl_Flora(list.tax, campañas, save = F, file = getwd())
}
\arguments{
\item{list.tax}{Listado taxonómico}

\item{campañas}{Campañas con parcelas de muestreo}

\item{save}{Logical. TRUE si desea guardar. FALSE default}

\item{file}{Directorio donde desea crear carpeta con resultados. Directorio de trabajo default.}
}
\value{
lista de resultados florísticos
}
\description{
\code{csl_Flora} arroja multiples resultados florísticos a partir del listado florístico y puntos de muestreo.
Es importante que los campos del listado taxonómico sean los mismos que se muestran en \code{Listado_taxonómico}.
Si utiliza el parámetro \code{save = T} entonces se creará una carpeta en el directorio que desee.
}
\examples{
csl_Flora(Listado_taxonómico,campañas)
csl_Flota(Listado_taxonómico,campañas, save = T)

}
