% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csl_FloraPotencial.R
\name{csl_FloraPotencial}
\alias{csl_FloraPotencial}
\title{FLora Potencial}
\usage{
csl_FloraPotencial(shape, link.simbio, save = FALSE, file)
}
\arguments{
\item{shape}{objeto sf.}

\item{link.simbio}{links de la o las páginas de los pisos vegetacionales correspondientes. Obtenerlos de \url{https://simbio.mma.gob.cl/Ecosistemas/}.}

\item{save}{Logical. Default False.}

\item{file}{Si save es TRUE. Ruta y nombre del archivo con terminal xlsx.}
}
\value{
flora potencial de acuerdo a Gajardo (1994) y Luebert y Pliscoff (2017)
}
\description{
FLora Potencial
}
\examples{
csl_FloraPotencial(shape = Pica, link.simbio = "https://simbio.mma.gob.cl/Ecosistemas/Details/51", save = T, file = "~/ruta/carpeta/file.xlsx")
}
