% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csl_kml_to_shp.R
\name{csl_kml_to_shp}
\alias{csl_kml_to_shp}
\title{Convertir archivos Kml a shp}
\usage{
csl_kml_to_shp(crs)
}
\arguments{
\item{crs}{Sistema de coordenadas de destino. Ej: 4326, 32718, 32719, etc.}
}
\value{
No retorna ni un archivo. Solo guarda shapefiles en el mismo directorio de los kml.
}
\description{
\code{csl_kml_to_shp} convierte los archivos kml de un directorio en shapefiles.
}
\examples{
csl_kml_to_shp(crs = 32719)

}
