% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csl_predio.R
\name{csl_predio}
\alias{csl_predio}
\title{Capas prediales para conaf}
\usage{
csl_predio(
  predio.shp,
  rol,
  Nombre.predio,
  save = F,
  NOMBREPREDIO,
  path = getwd()
)
}
\arguments{
\item{predio.shp}{Capa de un predio en formato sf y sistema de coordenadas UTM.}

\item{rol}{Rol de avalúo fiscal del predio}

\item{Nombre.predio}{Nombre del predio}

\item{save}{True si desea guardar la capa. FALSE default}

\item{NOMBREPREDIO}{Abreviación del nombre del predio que irá en el nombre del archivo al guardarse. Este debe ser todo con mayúsculas.}

\item{path}{Dirección donde desea guardar el documento. Directorio por defecto.}
}
\value{
Capa del predio como la solicita conaf
}
\description{
Capas prediales para conaf
}
\examples{
csl_predio(predio.shp = Pica,
rol = "123-4",
Nombre.predio = "Fulano y Sutano",
save = TRUE,
NOMBREPREDIO = "FULANOYSUTANO",
path = tempdir())

}
