% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pica.R
\docType{data}
\name{Pica}
\alias{Pica}
\title{Limites comuna de Pica, Chile}
\format{
\subsection{\code{Pica}}{

Simple feature collection with 1 feature and 7 fields:
\describe{
\item{CUT_REG, CUT_PROV, CUT_COM}{Código Únicos Territoriales}
\item{REGION}{Nombre de la región}
\item{PROVINCIA}{Nombre de la provincia}
\item{COMUNA}{Nombre de la comuna}
\item{SUPERFICIE}{superficie de la comuna en km2}
}
}
}
\source{
\url{https://www.ide.cl/index.php/limites-y-fronteras/item/1528-division-politica-administrativa-2020}
}
\usage{
Pica
}
\description{
sf de la comuna de Pica, con los límites e información politico-administrativa
}
\keyword{datasets}
